HTBHF-2672 Move request DTO objects into v2 folder.
