+++ api/src/main/java/uk/gov/dhsc/htbhf/claimant/converter/AddressToAddressDTOConverter.java
@@ -3,7 +3,7 @@ package uk.gov.dhsc.htbhf.claimant.converter;
 import org.springframework.stereotype.Component;
 import org.springframework.util.Assert;
 import uk.gov.dhsc.htbhf.claimant.entity.Address;
-import uk.gov.dhsc.htbhf.claimant.model.AddressDTO;
+import uk.gov.dhsc.htbhf.claimant.model.v2.AddressDTO;
 
 /**
  * Converts a {@link AddressDTO} into a {@link Address}.
