+++ api/src/main/java/uk/gov/dhsc/htbhf/claimant/converter/ClaimantDTOToClaimantConverter.java
@@ -3,7 +3,7 @@ package uk.gov.dhsc.htbhf.claimant.converter;
 import org.springframework.stereotype.Component;
 import org.springframework.util.Assert;
 import uk.gov.dhsc.htbhf.claimant.entity.Claimant;
-import uk.gov.dhsc.htbhf.claimant.model.ClaimantDTO;
+import uk.gov.dhsc.htbhf.claimant.model.v2.ClaimantDTO;
 
 import static org.apache.commons.lang3.StringUtils.upperCase;
