+++ api/src/main/java/uk/gov/dhsc/htbhf/claimant/model/eligibility/PersonDTO.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import uk.gov.dhsc.htbhf.claimant.model.AddressDTO;
+import uk.gov.dhsc.htbhf.claimant.model.v2.AddressDTO;
 
 import java.time.LocalDate;
