+++ api/src/test/java/uk/gov/dhsc/htbhf/claimant/model/AddressDTOTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import uk.gov.dhsc.htbhf.assertions.AbstractValidationTest;
+import uk.gov.dhsc.htbhf.claimant.model.v2.AddressDTO;
 
 import java.util.Set;
 import javax.validation.ConstraintViolation;
